+++ docs/HISTORY.rst
@@ -8,6 +8,7 @@ Change Log
 
 **Fixes**
 - Fixed issue whereby `turn_on`, `turn_off` and `toggle` for some HASS entities didn't work
+- Fixed issue whereby `render_template` didn't work
 
 **Breaking Changes**